split into two solutions, build out of source, update readme
