--- Samples/CppWinRTProjectionSample/SampleConsoleApp/Program.cs
@@ -1,25 +1,0 @@
-﻿//*********************************************************
-//
-// Copyright (c) Microsoft. All rights reserved.
-// This code is licensed under the MIT License (MIT).
-// THIS CODE IS PROVIDED *AS IS* WITHOUT WARRANTY OF
-// ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING ANY
-// IMPLIED WARRANTIES OF FITNESS FOR A PARTICULAR
-// PURPOSE, MERCHANTABILITY, OR NON-INFRINGEMENT.
-//
-//*********************************************************
-using System;
-
-namespace SampleConsoleApp
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var x = new SimpleMathComponent.SimpleMath();
-            Console.WriteLine("Adding 5.5 + 6.5 ...");
-            Console.WriteLine(x.add(5.5, 6.5).ToString());
-            Console.ReadLine();
-        }
-    }
-}
