--- Samples/CppWinRTProjectionSample/SampleConsoleApp/SampleConsoleApp.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net5.0-windows10.0.19041.0</TargetFramework>
-		<Platforms>x64</Platforms>
-	</PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.VCRTForwarders.140" Version="1.0.6" />
-    <PackageReference Include="SimpleMathComponent" Version="0.1.0-prerelease" />
-  </ItemGroup>
-
-</Project>
