--- Samples/CppWinRTProjectionSample/SimpleMathProjection/SimpleMathProjection.nuspec
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2012/06/nuspec.xsd">
-	<metadata>
-		<id>SimpleMathComponent</id>
-		<version>0.1.0-prerelease</version>
-		<authors>Contoso Math Inc.</authors>
-		<description>A simple component with basic math operations</description>
-		<dependencies>
-			<group targetFramework="net5.0">
-				<dependency id="Microsoft.Windows.CsWinRT" version="0.1.0-prerelease.200925.12" exclude="Build,Analyzers" />
-			</group>
-		</dependencies>
-	</metadata>
-	<files>
-		<!--Support net46+, netcore3, net5, uap, c++ -->
-		<file src="..\x64\Debug\SimpleMathComponent\SimpleMathComponent.winmd" target="lib\netcoreapp3.0\SimpleMathComponent.winmd" />
-		<file src="..\x64\Debug\SimpleMathComponent\SimpleMathComponent.winmd" target="lib\uap10.0\SimpleMathComponent.winmd" />
-		<file src="..\x64\Debug\SimpleMathComponent\SimpleMathComponent.winmd" target="lib\net46\SimpleMathComponent.winmd" />
-		<file src=".\bin\x64\Debug\net5.0-windows10.0.19041.0\SimpleMathProjection.dll" target="lib\net5.0\SimpleMathProjection.dll" />
-		<file src="..\x64\Debug\SimpleMathComponent\SimpleMathComponent.dll" target="runtimes\win10-x64\native\SimpleMathComponent.dll" />
-	</files>
-</package>
