+++ Samples/Net5ProjectionSample/ConsoleAppSample/ConsoleAppSample.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30509.190
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleAppSample", "ProjectionSampleConsoleApp\ConsoleAppSample.csproj", "{9623756E-90CE-449C-8BD8-36A2F8419DBF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Debug|x64.ActiveCfg = Debug|x64
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Debug|x64.Build.0 = Debug|x64
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Release|x64.ActiveCfg = Release|x64
+		{9623756E-90CE-449C-8BD8-36A2F8419DBF}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C799536A-3A0C-4EA8-8365-3CF29FA9F0D3}
+	EndGlobalSection
+EndGlobal
