+++ Samples/Net5ProjectionSample/ConsoleAppSample/ProjectionSampleConsoleApp/ConsoleAppSample.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0-windows10.0.19041.0</TargetFramework> 
+    <Platforms>x64</Platforms>
+  </PropertyGroup>
+
+  <PropertyGroup>
+	<RestoreSources>../../CppWinRTProjectionSample/SimpleMathProjection/nuget</RestoreSources>
+  </PropertyGroup>
+
+  <ItemGroup>
+	<PackageReference Include="Microsoft.VCRTForwarders.140" Version="1.0.6" />
+	<PackageReference Include="SimpleMathComponent" Version="0.1.0-prerelease" />
+  </ItemGroup>
+	
+</Project>
