+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/CppWinRTComponentProjectionSample.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.30404.54
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "SimpleMathComponent", "SimpleMathComponent\SimpleMathComponent.vcxproj", "{D6B0C16D-858A-4C1B-99CF-D6F4CF5BCD5F}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SampleConsoleApp", "SampleConsoleApp\SampleConsoleApp.csproj", "{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleMathProjection", "SimpleMathProjection\SimpleMathProjection.csproj", "{3D562B38-115F-481C-ADE2-5A60EFBA1ED0}"
 EndProject
 Global
@@ -41,26 +39,6 @@ Global
 		{D6B0C16D-858A-4C1B-99CF-D6F4CF5BCD5F}.Release|x64.Build.0 = Release|x64
 		{D6B0C16D-858A-4C1B-99CF-D6F4CF5BCD5F}.Release|x86.ActiveCfg = Release|Win32
 		{D6B0C16D-858A-4C1B-99CF-D6F4CF5BCD5F}.Release|x86.Build.0 = Release|Win32
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|ARM.Build.0 = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|ARM64.ActiveCfg = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|ARM64.Build.0 = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|x64.ActiveCfg = Debug|x64
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|x64.Build.0 = Debug|x64
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Debug|x86.Build.0 = Debug|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|ARM.ActiveCfg = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|ARM.Build.0 = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|ARM64.ActiveCfg = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|ARM64.Build.0 = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|x64.ActiveCfg = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|x64.Build.0 = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|x86.ActiveCfg = Release|Any CPU
-		{2DE91FD5-4584-47E4-9777-6BCBDEDCAAFD}.Release|x86.Build.0 = Release|Any CPU
 		{3D562B38-115F-481C-ADE2-5A60EFBA1ED0}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{3D562B38-115F-481C-ADE2-5A60EFBA1ED0}.Debug|ARM.ActiveCfg = Debug|x64
 		{3D562B38-115F-481C-ADE2-5A60EFBA1ED0}.Debug|ARM64.ActiveCfg = Debug|x64
