+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/SimpleMathComponent/PropertySheet.props
