+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/SimpleMathComponent/SimpleMath.idl
