+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/SimpleMathComponent/SimpleMathComponent.vcxproj.filters
