+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/SimpleMathComponent/packages.config
