+++ Samples/Net5ProjectionSample/CppWinRTProjectionSample/SimpleMathComponent/pch.cpp
