+++ Samples/Net5ProjectionSample/Directory.Build.targets
